requirements.txt

requests
Pillow
tkinter
